feat (user-api) - 유형별 적립 테이블의 내용에 따라 다른 포인트 값을 부여하도록 구현 #82

+++ src/main/java/com/nhnacademy/bookstoreuserapi/controller/PointTypeController.java
@@ -48,7 +48,7 @@ public class PointTypeController {
         return ResponseEntity.noContent().build();
     }
     @PutMapping("/{typeId}/point")
-    public ResponseEntity<ResponsePointType> updateEarningPoint(@PathVariable @NotNull @Min(1) Long typeId, @RequestParam @NotNull @Min(0) Long point) {
+    public ResponseEntity<ResponsePointType> updateEarningPoint(@PathVariable @NotNull @Min(1) Long typeId, @RequestParam @NotNull @Min(0) int point) {
 
         return ResponseEntity.ok().body(pointTypeService.updateEarningPoint(point, typeId));
     }
