+++ src/main/java/com/nhnacademy/bookstoreuserapi/domain/request/PointTypeCreateRequest.java
@@ -6,7 +6,7 @@ import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.Size;
 
 public record PointTypeCreateRequest (@NotBlank @Size(max = 20) String typeName,
-                                      @Min(0) Long earningPoint,
-                                      Integer earningRate,
+                                      @Min(0) int earningPoint,
+                                      int earningRate,
                                       @Size(max = 10) String gradeName){
 }
